# src/main/java/com/kazuya/teach_learning_backend/controller/AuthController.java
package com.kazuya.teach_learning_backend.controller;

import com.kazuya.teach_learning_backend.dto.LoginRequest;
import com.kazuya.teach_learning_backend.dto.LoginResponse;
import com.kazuya.teach_learning_backend.dto.SignupRequest;
import com.kazuya.teach_learning_backend.dto.SignupResponse;
import com.kazuya.teach_learning_backend.service.UserService;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;

@RestController
@RequestMapping("/api/auth")
@CrossOrigin(origins = "*") // フロント(React)から叩くなら一旦これでOK
public class AuthController {

    private final UserService userService;

    public AuthController(UserService userService) {
        this.userService = userService;
    }

    @PostMapping("/login")
    public ResponseEntity<?> login(@RequestBody LoginRequest request) {
        try {
            LoginResponse response = userService.login(request);
            return ResponseEntity.ok(response);
        } catch (RuntimeException e) {
            // 認証失敗時は 401 を返す
            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
                    .body(e.getMessage());
        }
    }

    @PostMapping("/signup")
    public ResponseEntity<?> signup(@RequestBody SignupRequest request) {
        try {
            SignupResponse response = userService.signup(request);
            return ResponseEntity.ok(response);
        } catch (RuntimeException e) {
            // 認証失敗時は 401 を返す
            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
                    .body(e.getMessage());
        }
    }

}